change type fields and add secondName to get fields echo ekg

+++ core/hsn/appointment/blocks/ekg/commands/create.py
@@ -25,15 +25,15 @@ class HsnCommandAppointmentBlockEkgCreateContext(BaseModel):
     trepetanie_predcerdiy: Optional[bool] = False
     another_changes: Optional[str] = None
     date_echo_ekg: str
-    fv: int
-    sdla: Optional[int] = None
-    lp: Optional[int] = None
-    pp: Optional[int] = None
-    kdr_lg: Optional[int] = None
-    ksr_lg: Optional[int] = None
-    kdo_lg: Optional[int] = None
-    mgp: Optional[int] = None
-    zslg: Optional[int] = None
+    fv: float
+    sdla: Optional[float] = None
+    lp: Optional[float] = None
+    pp: Optional[float] = None
+    kdr_lg: Optional[float] = None
+    ksr_lg: Optional[float] = None
+    kdo_lg: Optional[float] = None
+    mgp: Optional[float] = None
+    zslg: Optional[float] = None
 
     local_hypokines: Optional[bool] = False
     distol_disfunction: Optional[bool] = False
